Updated module for v0.2 of RC modular system
