+++ composer.json
@@ -13,7 +13,7 @@
     "prefer-stable": true,
     "require": {
         "php": "^5.4 | ^7.0",
-        "rebelcode/modular": "^0.1-alpha1",
+        "rebelcode/modular": "^0.2",
         "rebelcode/sql-expression-renderer": "^0.1-alpha1",
         "dhii/data-object-abstract": "^0.1-alpha1",
         "dhii/factory-base": "^0.1-alpha1"
@@ -23,6 +23,7 @@
         "ptrofimov/xpmock": "^1.1",
         "johnpbloch/wordpress-core": "^4.0",
         "dhii/php-cs-fixer-config": "^0.1",
+        "rebelcode/module-testing": "dev-task/initial-code",
         "codeclimate/php-test-reporter": "<=0.3.2",
         "dhii/data-container-interface": "dev-develop"
     },
