+++ module.php
@@ -3,13 +3,15 @@
 use Psr\Container\ContainerInterface;
 use RebelCode\Storage\Resource\WordPress\Module\WpCqrsModule;
 
-define('RC_WP_CQRS_MODULE_DIR', __DIR__);
-define('RC_WP_CQRS_MODULE_KEY', 'wp_cqrs');
-
 return function (ContainerInterface $c) {
     return new WpCqrsModule(
-        RC_WP_CQRS_MODULE_KEY,
-        [],
+        [
+            'key'                => 'wp_cqrs',
+            'dependencies'       => [],
+            'root_dir'           => __DIR__,
+            'config_file_path'   => __DIR__ . '/config.php',
+            'services_file_path' => __DIR__ . '/services.php',
+        ],
         $c->get('config_factory'),
         $c->get('container_factory'),
         $c->get('composite_container_factory')
